Add Object Oriented Programming classes
