+++ src/main/java/puke/Event.java
@@ -1,3 +1,5 @@
+package puke;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 public class Event extends Task {
