+++ src/main/java/puke/Task.java
@@ -1,3 +1,5 @@
+package puke;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
