+++ text-ui-test/EXPECTED.TXT
@@ -1,10 +1,17 @@
-Hello from puke
  ____         _
 |   | \ _   _| | _____
 |  _|  | | | | |/ / _ \
 | |    | |_| |   <  __/
 |_|    \__,_|_|\_\___|
 
+Features implemented and formats required:
+list: lists current tasks in list
+mark/unmark x : mark as done or unmark in list
+ToDo, Deadline, Event:
+todo xx
+deadline xx /by yyyy-mm-dd
+event xx /at yyyy-mm-dd
+delete x : remove the task from the list
     ____________________________________________________________
      OOPS!!! The Description of a ToDo cannot be empty.
     ____________________________________________________________
@@ -21,7 +28,7 @@ Hello from puke
     ____________________________________________________________
     ____________________________________________________________
      Got it. I've added this task:
-      [D][ ] ip (by: 5pm)
+      [D][ ] ip (by: Dec 1 2030)
      Now you have 2 tasks in the list.
     ____________________________________________________________
     ____________________________________________________________
@@ -31,7 +38,7 @@ Hello from puke
     ____________________________________________________________
      Here are the tasks in your list:
      1.[T][X] return book
-     2.[D][ ] ip (by: 5pm)
+     2.[D][ ] ip (by: Dec 1 2030)
     ____________________________________________________________
     ____________________________________________________________
      Noted. I've removed this task:
