+++ text-ui-test/pukeData.txt
@@ -1,0 +1,1 @@
+D | 0 | ip  | 2030-12-01
